Move components to safenet features folder
