+++ apps/web/src/components/settings/SettingsHeader/index.tsx
@@ -5,7 +5,7 @@ import PageHeader from '@/components/common/PageHeader'
 import css from '@/components/common/PageHeader/styles.module.css'
 import { generalSettingsNavItems, settingsNavItems } from '@/components/sidebar/SidebarNavigation/config'
 import { AppRoutes } from '@/config/routes'
-import { useHasSafenetFeature } from '@/features/safenet'
+import { useHasSafenetFeature } from '@/features/safenet/hooks/useHasSafenetFeature'
 import { useCurrentChain } from '@/hooks/useChains'
 import useSafeAddress from '@/hooks/useSafeAddress'
 import { isRouteEnabled } from '@/utils/chains'
