+++ apps/web/src/components/tx/SignOrExecuteForm/styles.module.css
@@ -75,11 +75,3 @@
   border-radius: 4px;
   padding: 2px 8px;
 }
-
-.safenetGradientCard {
-  margin-bottom: var(--space-2);
-}
-
-.safenetGradientCard > div:last-child > div:first-child {
-  margin-bottom: 0;
-}
