+++ apps/web/src/features/safenet/components/EnableSafenet/index.tsx
@@ -1,5 +1,5 @@
 import TxLayout from '@/components/tx-flow/common/TxLayout'
-import { ReviewEnableSafenet } from './ReviewEnableSafenet'
+import ReviewEnableSafenet from './ReviewEnableSafenet'
 
 export type EnableSafenetFlowProps = {
   guardAddress: string
