+++ apps/web/src/features/safenet/components/GradientBoxSafenet/index.tsx
