+++ apps/web/src/features/safenet/components/SafenetTxSimulation/styles.module.css
