+++ apps/web/src/features/safenet/hooks/useIsSafenetEnabled.ts
@@ -1,4 +1,4 @@
-import { useHasSafenetFeature } from '@/features/safenet'
+import { useHasSafenetFeature } from './useHasSafenetFeature'
 import useSafeInfo from '@/hooks/useSafeInfo'
 import { useGetSafenetConfigQuery } from '@/store/safenet'
 import { sameAddress } from '@/utils/addresses'
